backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "backend for background checker"
authors = ["Karthik Macherla <kmacher@seas.upenn.edu>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^3.2.2"
google-auth = "^2.9.1"
motor = "^3.0.0"
requests = "^2.28.1"
pydantic = {extras = ["email"], version = "^1.9.1"}
pymongo = "^4.2.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
yapf = "^0.32.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
